+++ gems2/mods/m_param.cpp
@@ -611,7 +611,7 @@ bool TProfil::rCopyFilterProfile( const char * prfName )
 // Modified on 15.11.2007 to return more detailed info on FIA and IPM iterations 
 // and precision refinement loops
 //
-double TProfil::calcMulti( int& NumPrecLoops, int& NumIterFIA, int& NumIterIPM )
+double TProfil::calcMulti( long int& NumPrecLoops, long int& NumIterFIA, long int& NumIterIPM )
 {
   TSysEq* STat = (TSysEq*)(&aMod[RT_SYSEQ]);
   calcFinished = false;
@@ -654,7 +654,7 @@ pmp->IT = pmp->ITG;   // This is to provide correct number of IPM iterations to 
 
    if( pa.p.PRD < 0 && pa.p.PRD > -50 ) // max 50 loops
    {  // Test refinement loops for highly non-ideal systems Added here by KD on 15.11.2007
-      int pp, pNPo = pmp->pNP,  TotW1 = pmp->W1+pmp->K2-1,  
+      long int pp, pNPo = pmp->pNP,  TotW1 = pmp->W1+pmp->K2-1,  
 ITstart = 10,        TotIT = pmp->IT;   // ITold = pmp->IT,
       pmp->pNP = 1;
       for( pp=0; pp < abs(pa.p.PRD); pp++ )
